backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.10.0
redis>=5.0.0
python-dotenv>=1.0.0
google-genai>=0.2.0
google-cloud-texttospeech>=2.16.0